Move tests from example app into main package.
